Convert storage-share to use service tracing.
This converts the storage-share to use the service trace provider.
